+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [HexaPDF::Layout::TextLayouter] to return a new
   [HexaPDF::Layout::TextLayouter::Result] structure when `#fit` is called that
   includes the `#draw` method
+* [HexaPDF::Layout::TextLayouter#fit] to require the height argument
 
 ## 0.7.0 - 2018-06-19